+++ pyincore/analyses/epfdamage/epfdamage.py
@@ -123,8 +123,8 @@ class EpfDamage(BaseAnalysis):
         fragility_key = self.get_parameter("fragility_key")
 
         fragility_set = dict()
-        fragility_set = self.fragilitysvc.match_inventory_object(self.get_input_dataset("dfr3_mapping_set"), epfs,
-                                                                 fragility_key)
+        fragility_set = self.fragilitysvc.match_inventory(self.get_input_dataset("dfr3_mapping_set"), epfs,
+                                                          fragility_key)
         epf_results = []
 
         # Converting list of epfs into a dictionary for ease of reference
@@ -135,7 +135,7 @@ class EpfDamage(BaseAnalysis):
         del list_epfs  # Clear as it's not needed anymore
 
         processed_epf = []
-        grouped_epfs = AnalysisUtil.group_by_demand_type_object(epfs, fragility_set)
+        grouped_epfs = AnalysisUtil.group_by_demand_type(epfs, fragility_set)
         for demand, grouped_epf_items in grouped_epfs.items():
             input_demand_type = demand[0]
             input_demand_units = demand[1]
@@ -204,9 +204,9 @@ class EpfDamage(BaseAnalysis):
             liq_fragility_key = self.get_parameter("liquefaction_fragility_key")
             if liq_fragility_key is None:
                 liq_fragility_key = self.DEFAULT_LIQ_FRAGILITY_KEY
-            liq_fragility_set = self.fragilitysvc.match_inventory_object(self.get_input_dataset("dfr3_mapping_set"),
-                                                                         epfs, liq_fragility_key)
-            grouped_liq_epfs = AnalysisUtil.group_by_demand_type_object(epfs, liq_fragility_set)
+            liq_fragility_set = self.fragilitysvc.match_inventory(self.get_input_dataset("dfr3_mapping_set"), epfs,
+                                                                  liq_fragility_key)
+            grouped_liq_epfs = AnalysisUtil.group_by_demand_type(epfs, liq_fragility_set)
 
             for liq_demand, grouped_liq_epf_items in grouped_liq_epfs.items():
                 liq_input_demand_type = liq_demand[0]