+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  in local file path [INCORe1-662](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-662)
 
 ### Changed
+- Replace old analyses util methods with new methods that use DFR3 Classes [INCORE1-685](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-685)
 - Refactoring EPF damage analyses to use local dfr3 classes and methods; adding corresponding test folder and
  tests [INCORE1-663](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-663) 
 - Refactoring pipeline damage with repair rate analyses to use local dfr3 classes and methods; adding corresponding